requirements.txt
botcity-framework-web
webdriver-manager
selenium==4.10.0
setuptools